moved continua script to visualisation package
